[STAGE] fix: type distribution design

+++ src/game/mod.rs
@@ -1,6 +1,6 @@
 use crate::{assets::RpgTextures, constants::*, control::*};
-use avian2d::prelude::*;
 use bevy::prelude::*;
+use item::IsItem;
 
 pub mod health;
 pub mod item;
@@ -15,37 +15,27 @@ pub fn inputs_use(
     actors: Query<Entity, With<Actor>>,
     q_st: Query<&item::ItemStorage>,
     q_pos: Query<&Transform>,
+    q_it: Query<&IsItem>,
     // asset_server: Res<AssetServer>,
     rpg_folder: Res<RpgTextures>,
 ) {
     let actor = actors.single();
     if click.just_pressed(MouseButton::Left) || control_settings.check(ControlCode::Use, &keys) {
         let storage = q_st.get(actor).unwrap();
-        let _item = if let Some(item) = storage.get_index(0) {
+        let item = if let Some(item) = storage.get_index(0) {
             item
         } else {
             return;
         };
 
-        let start_point =
-            q_pos.get(actor).unwrap().translation.truncate() + CHARACTER_LEFT_HAND_OFFSET;
-        let ray = coords.0.unwrap() - start_point;
-        let unit = ray / ray.length();
+        let it = q_it.get(item).unwrap();
+        if !it.0.check_can_use() {
+            return;
+        }
 
-        commands.spawn((
-            Sprite::from_atlas_image(
-                rpg_folder.get_image_handle("spells"),
-                rpg_folder.get_texture_atlas("spells", 3),
-            ),
-            Transform::from_xyz(
-                start_point.x + unit.x * 24.0,
-                start_point.y + unit.y * 24.0,
-                0.0,
-            ),
-            RigidBody::Dynamic,
-            Collider::circle(5.0),
-            LockedAxes::ROTATION_LOCKED,
-            LinearVelocity(unit * 40.0),
-        ));
+        let hold_point =
+            q_pos.get(actor).unwrap().translation.truncate() + CHARACTER_LEFT_HAND_OFFSET;
+
+        it.0.item_use(&mut commands, hold_point, coords, rpg_folder);
     }
 }
